Create Assignment 4 - Minor
This is my chatbot for TAG Directed Study. You can interact with it and have a short conversation. To begin talking, simply ask a question such as "how are you" or "what is your name". To exit the conversation, please say "goodbye". This project taught me how to use string manipulation in order to create a conversational program.